Survey-form-HTML-CSS
## Survey Form Project

This repository contains a survey form project designed to gather insights on [brief description of the survey topic]. The survey form aims to [state the objective of the survey].

### Key Features:
- User-friendly survey form interface.
- Comprehensive fields to collect relevant data effectively.
- Data validation and error-handling mechanisms for accuracy.
- Visually appealing design tailored to the survey's theme.

### Project Goals:
- Master HTML and CSS

Feel free to contribute to this project by providing feedback, suggestions, or enhancements. Your input is valuable in improving the survey form's functionality and usability.